feat: added reading time helper for blog, ui changes & refactored existing code related to the Blog.tsx component
