+++ src/components/Blog/Blog.tsx
@@ -1,5 +1,6 @@
 import type { Blog as BlogType } from '../../types';
-import getBlogContentPreview from './utils/getBlogContentPreview';
+import getBlogContentPreview from './helpers/getBlogContentPreview';
+import getReadingTime from './helpers/getReadingTime';
 import styles from './blog.module.scss';
 
 type BlogProps = Omit<BlogType, 'author'> & {
@@ -8,20 +9,21 @@ type BlogProps = Omit<BlogType, 'author'> & {
 
 export default function Blog(props: BlogProps) {
   const {
-    blogId,
     title,
     content,
     authorName,
   } = props;
 
   return (
-    <div
-      className={styles.blog}
-      key={blogId}
-    >
-      <h3>{title}</h3>
-      <p>{getBlogContentPreview(content)}</p>
-      <span>Written by {authorName}</span>
+    <div className={styles.blog}>
+      <span className={styles.blog__author}>
+        Written by <span>{authorName}</span>
+      </span>
+      <h2>{title}</h2>
+      <h3>{getBlogContentPreview(content)}</h3>
+      <span className={styles['blog__reading-time']}>
+        {getReadingTime(content)} min read
+      </span>
     </div>
   )
 }
