+++ src/components/Blog/blog.module.scss
@@ -1,3 +1,36 @@
 .blog {
   text-align: justify;
+  box-sizing: border-box;
+  padding: 1.25rem 0;
+
+  h2 {
+    color: rgba(41, 41, 41, 1);
+    font-size: 1.5rem;
+  }
+
+  h3 {
+    color: rgba(117, 117, 117, 1);
+    font-weight: 500;
+    font-size: 0.925rem;
+    margin-bottom: 0.275rem;
+  }
+
+  .blog__author {
+    color: rgba(117, 117, 117, 1);
+    font-weight: 400;
+
+    span {
+      font-weight: 600;
+      color: rgba(41, 41, 41, 1);
+    }
+  }
+
+  .blog__reading-time {
+    color: rgba(117, 117, 117, 1);
+    font-weight: 400;
+  }
+
+  span {
+    font-size: 0.85rem;
+  }
 }
